Added PlayerMovement Class
