+++ Assets/Static Prefabs/Player.prefab
@@ -219,6 +219,7 @@ GameObject:
   m_Component:
   - component: {fileID: 2941444580590092157}
   - component: {fileID: 2941444580590092156}
+  - component: {fileID: 8330837168031155109}
   - component: {fileID: 2941444580590092147}
   m_Layer: 0
   m_Name: Player
@@ -262,6 +263,26 @@ CharacterController:
   m_SkinWidth: 0.08
   m_MinMoveDistance: 0.001
   m_Center: {x: 0, y: -0.25, z: 0}
+--- !u!114 &8330837168031155109
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2941444580590092158}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c425fab5f3ffe2442a81e95700dd4bc4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  speed: 12
+  jumpHeight: 5
+  gravity: -9.81
+  groundDistance: 0.4
+  groundMask:
+    serializedVersion: 2
+    m_Bits: 256
+  groundCheck: {fileID: 2941444580297191603}
 --- !u!114 &2941444580590092147
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -278,11 +299,3 @@ MonoBehaviour:
     serializedVersion: 2
     m_Bits: 512
   focus: {fileID: 0}
-  speed: 12
-  jumpHeight: 2.5
-  gravity: -19.62
-  groundDistance: 0.4
-  groundMask:
-    serializedVersion: 2
-    m_Bits: 256
-  groundCheck: {fileID: 2941444580297191603}
